Feat : Server 나가기 기능 예외 처리 추가
서버 생성자는 해당서버에서 나갈 수 없도록 처리하였습니다.

+++ backend/src/user-server/user-server.service.ts
@@ -39,7 +39,16 @@ export class UserServerService {
     return this.userServerRepository.save(newUserServer);
   }
 
-  deleteById(id: number): Promise<DeleteResult> {
+  async deleteById(id: number, userId: number): Promise<DeleteResult> {
+    const userServer = await this.userServerRepository.findWithServerOwner(id);
+
+    if (!userServer) {
+      throw new BadRequestException('해당 서버에 참가하고 있지 않습니다.');
+    }
+    if (userServer.server.owner.id === userId) {
+      throw new BadRequestException('서버 생성자는 서버에서 나갈 수 없습니다.');
+    }
+
     return this.userServerRepository.delete(id);
   }
